requirements.txt
urequests
dht11
